Move input and output getters to native document
